Make the fatal warnings non fatal on release builds
The idea with fatal warning is a solid one: it is bad style to have code
compiling with warnings that could be easily fixed. But on the other
hand, having warnings enabled for release builds can cause issues like
this time with the deprecation of the Std module, which is not a
problem with code itself but rather with the environment moving forward.

Therefore this PR proposes to add a flag to the build which disables
fatal warning on release builds. This way the releases of the library
are forward compatible and will not be broken by future versions of
Core, Async or the OCaml compiler as easily.