Use the RELEASE flag consistently
As `install` calls `omake` it also needs to be passed the release
variable. Since `omake` is called in multiple places, it makes sense to
have its job control options defined in one place $(OMAKE) and the
number of parallel jobs can be gotten from the `nproc` executable and
default to the previous 4 if it is not available for some reason.