Import copy of AMQP spec into repository
This avoids the need of talking to the RabbitMQ servers every time the
library is built (thus avoiding failures if the server is down), which
also solves a problem when building with OPAM 2.0~rc2 with sandboxing
enabled since it prevents network access.

There is a Makefile target provided to update the specification
manually. This file has to be committed to git by hand but changes to
this file are probably exceedingly rare.

Closes #18.